moved settings to objects, show guides setting

+++ src/shaders/boundary.frag
@@ -1,4 +1,4 @@
-precision highp float;
+precision mediump float;
 
 uniform sampler2D u_readTexture;
 
@@ -11,6 +11,6 @@ void main( ) {
     vec3 pos = get3DFragCoord( );
 
     gl_FragColor = vec4( u_scale * texture3D( u_readTexture, (pos + v_offset) ).xyz, 1.0 );
-    //gl_FragColor = vec4( u_scale * v_offset, 1.0 );
+    //gl_FragColor = vec4( v_offset, 1.0 );
     //gl_FragColor = vec4( 0., 1.0, 0.5, 1.0);
 }
